refactor: rename event file to webhook_event
